+++ filament/backend/src/opengl/GLTexture.h
@@ -30,7 +30,7 @@ namespace filament::backend {
 struct GLTexture : public HwTexture {
     using HwTexture::HwTexture;
     struct GL {
-        GL() noexcept : imported(false), sidecarSamples(1), reserved(0) {}
+        GL() noexcept : imported(false), sidecarSamples(1), reserved1(0) {}
         GLuint id = 0;          // texture or renderbuffer id
         GLenum target = 0;
         GLenum internalFormat = 0;
@@ -40,10 +40,10 @@ struct GLTexture : public HwTexture {
         GLfloat anisotropy = 1.0;
         int8_t baseLevel = 127;
         int8_t maxLevel = -1;
-        uint8_t targetIndex = 0;    // optimization: index corresponding to target
+        uint8_t reserved0 = 0;
         bool imported           : 1;
         uint8_t sidecarSamples  : 4;
-        uint8_t reserved        : 3;
+        uint8_t reserved1       : 3;
     } gl;
 
     OpenGLPlatform::ExternalTexture* externalTexture = nullptr;