+++ filament/src/details/SwapChain.cpp
@@ -69,13 +69,14 @@ void FSwapChain::terminate(FEngine& engine) noexcept {
     engine.getDriverApi().destroySwapChain(mHwSwapChain);
 }
 
-void FSwapChain::setFrameScheduledCallback(FrameScheduledCallback callback, void* user) {
-    mFrameScheduledCallback = callback;
-    mEngine.getDriverApi().setFrameScheduledCallback(mHwSwapChain, callback, user);
+void FSwapChain::setFrameScheduledCallback(
+        backend::CallbackHandler* handler, FrameScheduledCallback&& callback) {
+    mFrameScheduledCallbackIsSet = bool(callback);
+    mEngine.getDriverApi().setFrameScheduledCallback(mHwSwapChain, handler, std::move(callback));
 }
 
-SwapChain::FrameScheduledCallback FSwapChain::getFrameScheduledCallback() const noexcept {
-    return mFrameScheduledCallback;
+bool FSwapChain::isFrameScheduledCallbackSet() const noexcept {
+    return mFrameScheduledCallbackIsSet;
 }
 
 void FSwapChain::setFrameCompletedCallback(backend::CallbackHandler* handler,